Updated code to examples and added activities controller-, model and routes.

+++ src/controllers/user-controller.mjs
@@ -71,10 +71,19 @@ const putUser = async (req, res) => {
   }
 };
 
-const deleteUser = async (req, res) => {
+const deleteUser = async (req, res, next) => {
+  // console.log('deleteUser', req.user, req.params.id);
+  // admin user can delete any user
+  // user authenticated by token can delete itself
+  if (
+    req.user.user_level !== 'admin' &&
+    req.user.user_id !== parseInt(req.params.id)
+  ) {
+    return next(customError('Unauthorized', 401));
+  }
   const result = await deleteUserById(req.params.id);
   if (result.error) {
-    return res.status(result.error).json(result);
+    return next(customError(result, result.error));
   }
   return res.json(result);
 };
